+++ shared/src/commonMain/kotlin/ui/viewmodels/authentication/AuthenticationSuccessViewModel.kt
@@ -1,3 +1,7 @@
 package ui.viewmodels.authentication
 
-class AuthenticationSuccessViewModel(val navigateUp: () -> Unit, val onClickLogo: () -> Unit)
+class AuthenticationSuccessViewModel(
+    val navigateUp: () -> Unit,
+    val onClickLogo: () -> Unit,
+    val openRedirectUrl: (() -> Unit)?,
+)