+++ shared/src/commonMain/kotlin/ui/views/Authentication/AuthenticationSuccessView.kt
@@ -27,6 +27,7 @@ import org.jetbrains.compose.resources.stringResource
 import ui.composables.Logo
 import ui.composables.buttons.ConcludeButton
 import ui.composables.buttons.NavigateUpButton
+import ui.composables.buttons.OpenUrlButton
 import ui.viewmodels.authentication.AuthenticationSuccessViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -62,9 +63,12 @@ fun AuthenticationSuccessView(
             ) {
                 Row(
                     modifier = Modifier.fillMaxWidth().padding(vertical = 16.dp),
-                    horizontalArrangement = Arrangement.Center,
+                    horizontalArrangement = Arrangement.SpaceEvenly,
                 ) {
                     ConcludeButton(vm.navigateUp)
+                    vm.openRedirectUrl?.let {
+                        OpenUrlButton(vm.openRedirectUrl)
+                    }
                 }
             }
         }